@@ -1,14 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 32
 
     defaultConfig {
         applicationId "org.mooncolony.moonmayor.captainsonarradarcompanion"
         minSdkVersion 16
         targetSdkVersion 32
-        versionCode 6
+        versionCode 7
         versionName "1.5"
     }
     buildTypes {
@@ -20,17 +19,14 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-
-    //butterknife dependencies
-    implementation 'com.jakewharton:butterknife:8.8.1'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    testImplementation 'junit:junit:4.12'
 
     //RecyclerView
-    compile group: 'com.android.support', name: 'recyclerview-v7', version: '25.3.0'
-    compile group: 'com.android.support', name: 'appcompat-v7', version: '25.3.0'
-    compile group: 'com.android.support', name: 'support-compat', version: '25.3.0'
-    compile group: 'com.android.support', name: 'animated-vector-drawable', version: '25.3.0'
-    compile group: 'com.android.support', name: 'design', version: '25.3.0'
+    implementation group: 'androidx.recyclerview', name: 'recyclerview', version: '1.0.0'
+    implementation group: 'androidx.appcompat', name: 'appcompat', version: '1.0.0'
+    implementation group: 'androidx.core', name: 'core', version: '1.0.0'
+    implementation group: 'androidx.vectordrawable', name: 'vectordrawable-animated', version: '1.0.0'
+    implementation group: 'com.google.android.material', name: 'material', version: '1.0.0'
 }
